WIP config after resize view
