+++ Wire-iOS.xcodeproj/xcshareddata/xcbaselines/BACB88501AF7C48900DDCDB0.xcbaseline/7267E61F-567A-490F-B29E-0F192958B13B.plist
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>classNames</key>
+	<dict>
+		<key>ContactsDataSourceTests</key>
+		<dict>
+			<key>testPerformanceExample</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.0039849</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>7. Dec 2018 at 13:47:35</string>
+				</dict>
+			</dict>
+		</dict>
+		<key>EmoticonSubstitutionConfigurationTests</key>
+		<dict>
+			<key>testThatParsingPerformanceForFullConfigurationIsEnoughForUsingOnMainQueue</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.00045093</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>7. Dec 2018 at 13:47:35</string>
+				</dict>
+			</dict>
+		</dict>
+	</dict>
+</dict>
+</plist>
